+++ SocialApp-SPA/package.json
@@ -27,7 +27,6 @@
     "core-js": "^2.5.4",
     "font-awesome": "^4.7.0",
     "ngx-bootstrap": "^3.0.1",
-    "ngx-gallery": "^5.6.2",
     "rxjs": "^6.0.0",
     "zone.js": "^0.8.26"
   },
