--- SocialApp-SPA/src/app/_models/photo.ts
@@ -1,7 +1,0 @@
-export interface Photo {
-    id:number;
-    url:string;
-    description:string;
-    dateAdded:Date;
-    isMain:boolean;
-}
