--- SocialApp-SPA/src/app/_resolver/member-detail.resolver.ts
@@ -1,22 +1,0 @@
-import { Injectable } from "@angular/core";
-import { Resolve, Route, Router, ActivatedRoute, ActivatedRouteSnapshot } from "@angular/router";
-import { User } from "../_models/user";
-import { UserService } from "../_services/user.service";
-import { AlertifyService } from "../_services/alertify.service";
-import { Observable, of } from "rxjs";
-import { catchError } from "rxjs/operators";
-
-@Injectable()
-export class MemberDetailResolver implements Resolve<User> {
-    constructor(private userService:UserService,private router:Router,private alertify:AlertifyService){}
-
-    resolve(route:ActivatedRouteSnapshot):Observable<User>{
-        return this.userService.getUser(route.params['id']).pipe(
-            catchError(error=>{
-                this.alertify.error('Problem retrieving data');
-                this.router.navigate(['/members']);
-                return of(null);
-            })
-        );
-    }
-}
