+++ SocialApp-SPA/src/app/app.module.ts
@@ -5,32 +5,18 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import {HttpClientModule} from '@angular/common/http';
 import {FormsModule} from '@angular/forms';
-import { BsDropdownModule, TabsModule } from 'ngx-bootstrap';
+import { BsDropdownModule } from 'ngx-bootstrap';
 import { RouterModule } from '../../node_modules/@angular/router';
-import { JwtModule } from '@auth0/angular-jwt';
-import { NgxGalleryModule } from 'ngx-gallery';
-
 
 import { AppComponent } from './app.component';
 import { NavComponent } from './nav/nav.component';
 import { RegisterComponent } from './register/register.component';
 import { HomeComponent } from './home/home.component';
-import { MemberListComponent } from './members/member-list/member-list.component';
+import { MemberListComponent } from './member-list/member-list.component';
 import { ListsComponent } from './lists/lists.component';
 import { MessagesComponent } from './messages/messages.component';
 import { appRoutes } from './routes';
 import { AuthGuard } from './_guards/auth.guard';
-import { UserService } from './_services/user.service';
-import { MemberCardComponent } from './members/member-card/member-card.component';
-import { MemberDetailComponent } from './members/member-detail/member-detail.component';
-import { MemberDetailResolver } from './_resolver/member-detail.resolver';
-import { MemberListResolver } from './_resolver/member-list.resolver';
-
-
-export function tokenGetter(){
-    return localStorage.getItem('token');
-}
-
 @NgModule({
    declarations: [
       AppComponent,
@@ -39,34 +25,20 @@ export function tokenGetter(){
       HomeComponent,
       MemberListComponent,
       ListsComponent,
-      MessagesComponent,
-      MemberCardComponent,
-      MemberDetailComponent
+      MessagesComponent
    ],
    imports: [
       BrowserModule,
       HttpClientModule,
       FormsModule,
       BsDropdownModule.forRoot(),
-      TabsModule.forRoot(),
-      RouterModule.forRoot(appRoutes),
-      NgxGalleryModule,
-      JwtModule.forRoot({
-          config:{
-              tokenGetter:tokenGetter,
-              whitelistedDomains:['localhost:5000'],
-              blacklistedRoutes:['localhost:5000/api/auth']
-          }
-      })
+      RouterModule.forRoot(appRoutes)
    ],
    providers: [
       AuthService,
       ErrorInterceptorProvider,
       AlertifyService,
-      AuthGuard,
-      UserService,
-      MemberDetailResolver,
-      MemberListResolver
+      AuthGuard
    ],
    bootstrap: [
       AppComponent
