+++ SocialApp-SPA/src/app/member-list/member-list.component.html
@@ -1,0 +1,3 @@
+<p>
+  member-list works!
+</p>
