--- SocialApp-SPA/src/app/members/member-card/member-card.component.css
@@ -1,36 +1,0 @@
-.card:hover img{
-    transform: scale(1.2,1.2);
-    transition-duration: 500ms;
-    transition-timing-function: ease-out;
-    opacity: 0.7;
-}
-
-.card img{
-    transform: scale(1.0,1.0);
-    transition-duration: 500ms;
-    transition-timing-function: ease-out;
-}
-
-.card-img-wrapper{
-    overflow: hidden;
-    position: relative;
-}
-
-.member-icons{
-    position: absolute;
-    bottom: -30%;
-    left: 0;
-    right: 0;
-    margin-right: auto;
-    margin-left: auto;
-    opacity: 0;
-}
-
-.card-img-wrapper:hover .member-icons{
-    bottom: 0;
-    opacity: 1;
-}
-
-.animate{
-    transition:all 0.3s ease-in-out;
-}
