--- SocialApp-SPA/src/app/members/member-card/member-card.component.html
@@ -1,20 +1,0 @@
-<div class="card mb-4 mt-5">
-  <div class="card-img-wrapper">
-    <img src="{{user.photoUrl}}" alt="{{user.knownAs}}" class="card-img-top">
-    <ul class="list-inline member-icons animate text-center">
-      <li class="list-inline-item">
-        <button class="btn btn-primary" [routerLink]="['/members/',user.id]"><i class="fa fa-user"></i></button>
-      </li>
-      <li class="list-inline-item">
-        <button class="btn btn-primary"><i class="fa fa-heart"></i></button>
-      </li>
-      <li class="list-inline-item">
-        <button class="btn btn-primary"><i class="fa fa-envelope"></i></button>
-      </li>
-    </ul>
-  </div>
-  <h6 class="card-title text-center mb-1">
-    <i class="fa fa-user"> {{user.knownAs}}, {{user.age}}</i>
-    <p class="card-text text-muted text-center">{{user.city}}</p>
-  </h6>
-</div>
