--- SocialApp-SPA/src/app/members/member-detail/member-detail.component.css
@@ -1,16 +1,0 @@
-.img-thumbnail{
-    margin: 25px;
-    width: 85%;
-    height: 85%;
-}
-
-.card-body{
-    padding: 0 25px;
-}
-
-.card-footer{
-    padding: 10px 15px;
-    background-color: #fff;
-    border-top: none;
-}
-
