--- SocialApp-SPA/src/app/members/member-detail/member-detail.component.html
@@ -1,60 +1,0 @@
-<div class="container mt-4">
-  <div class="row">
-    <h1>{{user.knownAs}}'s Profile</h1>
-  </div>
-  <div class="row">
-    <div class="col-sm-4">
-      <div class="card">
-        <img class="card-img-top img-thumbnail" src="{{user.photoUrl}}" alt="{{user.knownAs}}">
-        <div class="card-body">
-          <div>
-            <strong>Location:</strong>
-            <p>{{user.city}}, {{user.country}}</p>
-          </div>
-          <div>
-            <strong>Age:</strong>
-            <p>{{user.age}}</p>
-          </div>
-          <div>
-            <strong>Last Active:</strong>
-            <p>{{user.lastActive}}</p>
-          </div>
-          <div>
-            <strong>Member since:</strong>
-            <p>{{user.created}}</p>
-          </div>
-          <div class="car-footer">
-            <div class="btn-group d-flex">
-              <button class="btn btn-primary w-100">Like</button>
-              <button class="btn btn-success w-100">Message</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-sm-8">
-      <div class="tab-panel">
-        <tabset>
-          <tab heading="About {{user?.knownAs}}">
-            <h4>Description</h4>
-            <p>{{user.introduction}}</p>
-            <h4>Looking For</h4>
-            <p>{{user.lookingFor}}</p>
-          </tab>
-          <tab heading="Interests">
-            <h4>Interests</h4>
-            <p>{{user.interests}}</p>
-          </tab>
-          <tab heading="Photos">
-              <h4>Photos</h4>
-              <ngx-gallery [options]="galleryOptions" [images]="galleryImages"></ngx-gallery>
-            </tab>
-            <tab heading="Messages">
-                <h4>Messages</h4>
-                <p>Messages will go here</p>
-              </tab>
-        </tabset>
-      </div>
-    </div>
-  </div>
-</div>
