--- SocialApp-SPA/src/app/members/member-list/member-list.component.html
@@ -1,7 +1,0 @@
-<div class="container">
-  <div class="row">
-    <div *ngFor="let user of users" class="col-lg-2 col-md-3 col-sm-6">
-      <app-member-card [user]="user"></app-member-card>
-    </div>
-  </div>
-</div>
