--- SocialApp-SPA/src/app/members/member-list/member-list.component.ts
@@ -1,31 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { User } from '../../_models/user';
-import { AlertifyService } from '../../_services/alertify.service';
-import { UserService } from '../../_services/user.service';
-import { ActivatedRoute } from '@angular/router';
-
-@Component({
-  selector: 'app-member-list',
-  templateUrl: './member-list.component.html',
-  styleUrls: ['./member-list.component.css']
-})
-export class MemberListComponent implements OnInit {
-  users:User[];
-
-  constructor(private userService:UserService,private alertify:AlertifyService,private route:ActivatedRoute) { }
-
-  ngOnInit() {
-    this.route.data.subscribe(data=>{
-      this.users=data['users'];
-    });
-  }
-
-  // getUsers(){
-  //   this.userService.getUsers().subscribe((users:User[])=>{
-  //     this.users=users;
-  //   },error=>{
-  //     this.alertify.error(error);
-  //   });
-  // }
-
-}
