+++ SocialApp-SPA/src/app/routes.ts
@@ -1,12 +1,9 @@
 import { ListsComponent } from './lists/lists.component';
 import { MessagesComponent } from './messages/messages.component';
-import { MemberListComponent } from './members/member-list/member-list.component';
+import { MemberListComponent } from './member-list/member-list.component';
 import { HomeComponent } from './home/home.component';
 import { Routes } from "@angular/router";
 import { AuthGuard } from './_guards/auth.guard';
-import { MemberDetailComponent } from './members/member-detail/member-detail.component';
-import { MemberDetailResolver } from './_resolver/member-detail.resolver';
-import { MemberListResolver } from './_resolver/member-list.resolver';
 
 export const appRoutes:Routes=[
     {path:'',component:HomeComponent},
@@ -15,8 +12,7 @@ export const appRoutes:Routes=[
         runGuardsAndResolvers:'always',
         canActivate:[AuthGuard],
         children:[
-            {path:'members',component:MemberListComponent,resolve:{users:MemberListResolver}},
-            {path:'members/:id',component:MemberDetailComponent,resolve:{user:MemberDetailResolver}},
+            {path:'members',component:MemberListComponent},
             {path:'messages',component:MessagesComponent},
             {path:'lists',component:ListsComponent}
         ]
