+++ SocialApp-SPA/src/environments/environment.ts
@@ -3,9 +3,7 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  apiUrl:'http://localhost:5000/api/',
-  hmr:true
+  production: false
 };
 
 /*
